swift/requirements.txt
noise==1.2.2
numpy==1.26.4
pillow==8.4.0
pyshp==2.3.1
scipy==1.13.0
